fix: update ReportDto to use double and adjust CalculateHoursWorked method accordignly

See merge request epm-edai/project-runs/run-7/team-2/microservices/restaurants!44

+++ Restaurant.Application/DTOs/Reports/ReportDto.cs
@@ -10,7 +10,7 @@
 
         public required string WaiterEmail { get; set; }
 
-        public required int HoursWorked { get; set; }
+        public required decimal HoursWorked { get; set; }
 
         public required double AverageServiceFeedback { get; set; }
